Lift restrictions about column ordering
It was based on the wrong assumption that the requiredColumns
that are pased to the buildScan functions are in the same order
as the selected columns of the Dataset.
This is not the case, and we can do without that restriction if we
make sure that the returned data has the same column ordering as
requiredColumns.